Adds module generate-ads and parametr-ads
