+++ js/generates-ads.js
@@ -1,59 +1,25 @@
 import {getRandomElementArr,  collectRandomEl,  getRandomFloat, getRandomInt} from './utils.js';
-
-//Parameters
-const TYPES = [
-  'palace',
-  'flat',
-  'house',
-  'bungalow',
-  'hotel'];
-
-const CHECKIN_TIMES = [
-  '12:00',
-  '13:00',
-  '14:00'
-];
-
-const FEATURES = [
-  'wifi',
-  'dishwasher',
-  'parking',
-  'washer',
-  'elevator',
-  'conditioner'
-];
-
-const PHOTOS = [
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
-  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
-];
-
-const Coordinates = {
-  LAT_MIN: 35.65000,
-  LAT_MAX: 35.70000,
-  LNG_MIN: 139.70000,
-  LNG_MAX: 139.80000
-};
+import {TYPES, CHECKIN_TIMES, FEATURES, PHOTOS, Coordinates} from './parameters-ads';
 
 //Количество выводимых объявлений
 const COUNT_ADS = 10;
 
-//Функция создает объявление
-const createsAds = (index) => {
+//Функция генерирует объявление
+
+function createsAds(index) {
   const location = {
     lat: getRandomFloat(Coordinates.LAT_MIN, Coordinates.LAT_MAX),
     lng: getRandomFloat(Coordinates.LNG_MIN, Coordinates.LNG_MAX),
   };
   return {
     author: {
-      avatar: `img/avatars/user${String(index).padStart(2,'0')}.png`
+      avatar: `img/avatars/user${String(index).padStart(2, '0')}.png`
     },
     offer: {
       title: 'Гостевой дом Бристоль',
       address: `${location.lat}, ${location.lng}`,
-      price: getRandomInt(5,10),
-      type : getRandomElementArr(TYPES),
+      price: getRandomInt(5, 10),
+      type: getRandomElementArr(TYPES),
       rooms: getRandomInt(1, 4),
       guests: getRandomInt(1, 6),
       checkin: getRandomElementArr(CHECKIN_TIMES),
@@ -64,7 +30,7 @@ const createsAds = (index) => {
     },
     location
   };
-};
+}
 
 const greateAdsArray = (count) => Array.from({length: count, createsAds}, (_, i) => createsAds(i + 1));
 greateAdsArray(COUNT_ADS);
